// sdk/src/models/message.ts
export type MessageThreadRootInfo = {
  replyCount?: number | undefined;
  lastReplyTimestampMilli?: number | undefined;
};

export type MessageThreadReplyInfo = {
  rootId?: string | undefined;
  replySentToChannel?: boolean | undefined;
  rootText?: string | undefined;
};

export type Message = {
  id: string;
  channelId: string;
  workspaceId: string;
  /**
   * ID of the user (or addon) that sent the message.
   */
  author: string;
  /**
   * ID of the addon app that authored the message (null for human-authored).
   */
  authorAppId?: string | null | undefined;
  text: string;
  blocks?: Array<MessageBlock> | null | undefined;
  attachments?: Array<{ [k: string]: any }> | undefined;
  files?: Array<{ [k: string]: any }> | undefined;
  reactions?: Array<Reaction> | undefined;
  linkPreviews?: Array<{ [k: string]: any }> | undefined;
  threadRootInfo?: MessageThreadRootInfo | null | undefined;
  threadReplyInfo?: MessageThreadReplyInfo | null | undefined;
  deleted?: boolean | undefined;
  edited?: boolean | undefined;
  isFollowing?: boolean | undefined;
  /**
   * Message subtype (empty string for normal messages).
   */
  subtype?: string | undefined;
  localId?: string | undefined;
  meta?: { [k: string]: any } | null | undefined;
  savedTimestampMilli?: number | undefined;
  timestamp: Date;
  timestampMilli: number;
};