// sdk/src/models/operations/fetch-thread-replies.ts
/*
 * Code generated by Speakeasy (https://speakeasy.com). DO NOT EDIT.
 */

import * as z from "zod/v3";
import { remap as remap$ } from "../../lib/primitives.js";
import { safeParse } from "../../lib/schemas.js";
import { Result as SafeParseResult } from "../../types/fp.js";
import * as types from "../../types/primitives.js";
import { SDKValidationError } from "../errors/sdk-validation-error.js";
import * as models from "../index.js";

export type FetchThreadRepliesRequest = {
  rootMessageId: string;
  channelId?: string | undefined;
  channel?: string | undefined;
  /**
   * ID of the last reply on the previous page.
   */
  cursor?: string | undefined;
  limit?: number | undefined;
};

export type FetchThreadRepliesResponse = {
  result: Array<models.Message>;
};

/** @internal */
export const FetchThreadRepliesRequest$inboundSchema: z.ZodType<
  FetchThreadRepliesRequest,
  z.ZodTypeDef,
  unknown
> = z.object({
  rootMessageId: types.string(),
  channelId: types.optional(types.string()),
  channel: types.optional(types.string()),
  cursor: types.optional(types.string()),
  limit: types.number().default(100),
});
/** @internal */
export type FetchThreadRepliesRequest$Outbound = {
  rootMessageId: string;
  channelId?: string | undefined;
  channel?: string | undefined;
  cursor?: string | undefined;
  limit: number;
};

/** @internal */
export const FetchThreadRepliesRequest$outboundSchema: z.ZodType<
  FetchThreadRepliesRequest$Outbound,
  z.ZodTypeDef,
  FetchThreadRepliesRequest
> = z.object({
  rootMessageId: z.string(),
  channelId: z.string().optional(),
  channel: z.string().optional(),
  cursor: z.string().optional(),
  limit: z.number().int().default(100),
});

export function fetchThreadRepliesRequestToJSON(
  fetchThreadRepliesRequest: FetchThreadRepliesRequest,
): string {
  return JSON.stringify(
    FetchThreadRepliesRequest$outboundSchema.parse(fetchThreadRepliesRequest),
  );
}
export function fetchThreadRepliesRequestFromJSON(
  jsonString: string,
): SafeParseResult<FetchThreadRepliesRequest, SDKValidationError> {
  return safeParse(
    jsonString,
    (x) => FetchThreadRepliesRequest$inboundSchema.parse(JSON.parse(x)),
    `Failed to parse 'FetchThreadRepliesRequest' from JSON`,
  );
}

/** @internal */
export const FetchThreadRepliesResponse$inboundSchema: z.ZodType<
  FetchThreadRepliesResponse,
  z.ZodTypeDef,
  unknown
> = z.object({
  Result: z.array(models.Message$inboundSchema),
}).transform((v) => {
  return remap$(v, {
    "Result": "result",
  });
});
/** @internal */
export type FetchThreadRepliesResponse$Outbound = {
  Result: Array<models.Message$Outbound>;
};

/** @internal */
export const FetchThreadRepliesResponse$outboundSchema: z.ZodType<
  FetchThreadRepliesResponse$Outbound,
  z.ZodTypeDef,
  FetchThreadRepliesResponse
> = z.object({
  result: z.array(models.Message$outboundSchema),
}).transform((v) => {
  return remap$(v, {
    result: "Result",
  });
});

export function fetchThreadRepliesResponseToJSON(
  fetchThreadRepliesResponse: FetchThreadRepliesResponse,
): string {
  return JSON.stringify(
    FetchThreadRepliesResponse$outboundSchema.parse(fetchThreadRepliesResponse),
  );
}
export function fetchThreadRepliesResponseFromJSON(
  jsonString: string,
): SafeParseResult<FetchThreadRepliesResponse, SDKValidationError> {
  return safeParse(
    jsonString,
    (x) => FetchThreadRepliesResponse$inboundSchema.parse(JSON.parse(x)),
    `Failed to parse 'FetchThreadRepliesResponse' from JSON`,
  );
}